Tradestation/requirements.txt
# TradeStation API Integration Dependencies

# Core HTTP client for API requests
requests>=2.28.0

# Environment variable management (optional but recommended)
python-dotenv>=1.0.0

# Logging and utilities (usually included with Python)
# logging - built-in
# datetime - built-in
# typing - built-in (Python 3.5+)
# dataclasses - built-in (Python 3.7+)
# pathlib - built-in (Python 3.4+)
# json - built-in
# os - built-in
# sys - built-in

# Optional: For enhanced functionality
# pandas>=1.5.0  # For data analysis
# numpy>=1.21.0  # For numerical operations
# matplotlib>=3.5.0  # For charting
# plotly>=5.0.0  # For interactive charts
